codes/hack-llm/pyproject.toml
[project]
name = "hack-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.8.5",
    "ddgs>=9.14.2",
    "gradio>=6.14.0",
    "httpx>=0.28.1",
    "huggingface-hub>=1.14.0",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "jupyter>=1.1.1",
    "langchain>=1.2.17",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.2",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.2.1",
    "langchain-postgres>=0.0.17",
    "langchain-text-splitters>=1.1.2",
    "langgraph>=1.1.10",
    "llama-index>=0.14.21",
    "llama-index-embeddings-huggingface>=0.7.0",
    "matplotlib>=3.10.9",
    "mcp[cli]>=1.27.0",
    "ollama>=0.6.2",
    "pandas>=3.0.2",
    "pip>=26.1.1",
    "pymupdf>=1.27.2.3",
    "python-dotenv>=1.2.2",
    "scikit-image>=0.26.0",
    "scikit-learn>=1.8.0",
    "smolagents[litellm]>=1.22.0",
    "timm>=1.0.26",
    "torch>=2.11.0",
    "transformers[serving,torch]>=5.8.0",
]

[dependency-groups]
dev = [
    "langgraph-cli>=0.4.26",
]
